Superficial changes, cleanup.
Renamed addToGraph_* to define.
Added more concise description to each MSVC build rule.

+++ Test2/clean.bat
@@ -1,3 +1,3 @@
 @echo off
-rmdir /S /Q Built
+if EXIST Built  rmdir /S /Q Built
